requirements.txt
Django
djangorestframework
# https://www.django-rest-framework.org/api-guide/authentication/#json-web-token-authentication
djangorestframework-simplejwt
django-filter
drf-nested-routers
